V2 and final
I figure i stick out with the logo of the school on the bottom side of the window because with image starts to look very ugly and the text label + buttons have a specific bgcolor of white and it interfears with the background image every  single time. I've tried different ways but this is the best one taking in ecuation how much time people will spend with this open.